move Bson <-> Data translation to separate file
Because the Bson type is part of our generic MongoDB wrapper, but Data
is not.
